adapt the test
